requirements.txt
flask
flask-cors
opencv-python-headless
numpy
ultralytics
Pillow
google-generativeai
python-dotenv
transformers
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.0.1+cpu
torchvision==0.15.2+cpu
gunicorn==20.1.0
Werkzeug>=2.0.0